+++ sitepress-rails/spec/dummy/app/controllers/secondary_controller.rb
@@ -1,0 +1,3 @@
+class SecondaryController < Sitepress::SiteController
+  self.site = Sitepress.sites.fetch(Rails.root.join("app/sitepress/secondary").to_s)
+end